requirement.txt
sklearn
matplotlib
pysam
pandas
numpy
pytorch
torchvision
